requirements.txt
fastapi==0.115.6
passlib==1.7.4
pydantic==2.10.4
pytest==8.3.4
python_jose==3.3.0
SQLAlchemy==2.0.36
